tests uint16 to uint8

+++ webnfc/scripts/main.js
@@ -47,7 +47,7 @@ function readNdefTag() {
 
 // https://developers.google.com/web/updates/2012/06/How-to-convert-ArrayBuffer-to-and-from-String
 function ab2str(buf) {
-  return String.fromCharCode.apply(null, new Uint16Array(buf));
+  return String.fromCharCode.apply(null, new Uint8Array(buf));
 }
